Fixed Github OAuth return not working and user info being incomplete
